% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_risk.R
\name{temp_risk}
\alias{temp_risk}
\title{temperature risk}
\usage{
temp_risk(
  temp,
  consecutive_days,
  temp_threshold = 30,
  trees,
  income,
  buildings
)
}
\arguments{
\item{temp}{(C) (at least 5 days)}

\item{consecutive_days}{length of the heat event}

\item{temp_threshold}{(C) (default 30)}

\item{trees}{number of trees in the area}

\item{income}{yearly income of the individual}

\item{buildings}{number of buildings in the area}
}
\value{
risk (high, med, low)
}
\description{
compute temperature risk for urban dwellers
}
